[refactor|test] Replace file paths
