+++ include/pbdt/bdd.hpp
@@ -22,7 +22,7 @@
 
 #endif
 
-#include "prop-pp/test-context.hpp"
+#include "pbdt/test-context.hpp"
 
 namespace pbdt::bdd
 {
