+++ tests/src/scenario.cpp
@@ -6,7 +6,7 @@
 
 #include <print>
 
-#include "prop-pp/bdd.hpp"
+#include "pbdt/bdd.hpp"
 
 #include "fixtures.hpp"
