+++ tests/src/test-context.cpp
@@ -6,7 +6,7 @@
 
 #include <print>
 
-#include "prop-pp/test-context.hpp"
+#include "pbdt/test-context.hpp"
 
 #include "fixtures.hpp"
