+++ tests/src/then-component.cpp
@@ -6,7 +6,7 @@
 
 #include <tuple>
 
-#include "prop-pp/bdd.hpp"
+#include "pbdt/bdd.hpp"
 
 #include "fixtures.hpp"
